+++ src/Controller/CurrencyController.php
@@ -7,6 +7,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\Form\Extension\Core\Type\IntegerType;
 use Symfony\Component\Form\Extension\Core\Type\SubmitType;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
+use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
@@ -65,7 +66,13 @@ class CurrencyController extends AbstractController
      * @Route("/currency/{id}", name="currency_edit", methods={"GET", "POST"})
      */
     public function edit(Request $request, int $id) {
-        $currency = $this->getDoctrine()->getRepository(Currency::class)->find($id);
+        $entityManager = $this->getDoctrine()->getManager();
+        $currency = $entityManager->find(Currency::class, $id);
+
+        if ($currency == null)
+        {
+            return new JsonResponse(['message' => "Currency not found"], 404);
+        }
 
         $form = $this->buildFormFields($currency)
             ->add('save', SubmitType::class, ['label' => 'Store', 'attr' => ['class' => 'btn btn-primary']])
@@ -76,7 +83,6 @@ class CurrencyController extends AbstractController
         if($form->isSubmitted() && $form->isValid()) {
             $country = $form->getData();
 
-            $entityManager = $this->getDoctrine()->getManager();
             $entityManager->persist($country);
             $entityManager->flush();
 
@@ -90,13 +96,23 @@ class CurrencyController extends AbstractController
      * @Route("/currency/{id}", name="currency_delete", methods={"DELETE"})
      */
     public function delete(Request $request, int $id) {
-        $currency = $this->getDoctrine()->getRepository(Currency::class)->find($id);
-  
         $entityManager = $this->getDoctrine()->getManager();
-        $entityManager->remove($currency);
-        $entityManager->flush();
-  
-        $response = new Response();
-        $response->send();
+        $currency = $entityManager->find(Currency::class, $id);
+
+        if ($currency == null)
+        {
+            return new JsonResponse(['message' => "Currency not found"], 404);
+        }
+
+        try
+        {
+            $entityManager->remove($currency);
+            $entityManager->flush();
+            return new JsonResponse(['message' => 'ok']);
+        }
+        catch (\Exception $e)
+        {
+            return new JsonResponse(['message' => $e->getMessage()], 409);
+        }
     }
 }
